requirements.txt
fastapi
uvicorn
streamlit
httpx
pydantic
pyyaml
